mfp/requirements.txt
Django==1.11.7
djangorestframework==3.8.2
django-filter==1.1.0
django-filters==0.2.1
celery==4.1.0
django-celery-results==1.0.1
